chore(settings): split device settings in multiple components to make it easier to read and edit
